requirements.txt
# Core LangChain
langchain
langchain-groq
langchain-core
langchain-community

# Environment variables
python-dotenv

# UI
streamlit

# PDF reading (Day 3)
pypdf
faiss-cpu

sentence-transformers
langchain-huggingface
langchain-text-splitters